NTBS-414 fix tests and apply new validation to nhs number
